Add 800+ common words in one click
Updated the app offline dictionary, added list of common words, fixed some problems in the search function, addded tests and fixed a few bugs. The search now only looks for words that start with what you typed. The search is now cached and should be faster when typing long words.